implement reading/writing of MC data
Input file changes:
- additional line for the filename of the MC data
- last line is now the job type: MC or PFE/Exact

For MC job, only MC sampling is run and data is saved.

For other job, MC data is read in, and the given method
is used to estimate the partition function.

When reading MC data, consistency with the current system
parameters is checked. If the MC data contains more samples
than needed by the input, an informational message is printed,
and only the required part of the MC data is read in.