PartFunc3: only iterate to Estar
We first find Estar like in PartFunc (but with a starting point
that is not at Emax) and then perform the NSVolume iteration only
for levels >= Estar, recording the total error as we go along.
This avoids doing the volume iteration further then necessary.

Also, transplant the movedist statistics into PartFunc3.